to raise precision switch content value type from float to double

+++ src/UnitConv/IUnitConverter.cs
@@ -6,6 +6,6 @@ namespace UnitConv
 {
     public interface IUnitConverter
     {
-        Length ConvertLength(float value, LengthUnit fromUnit, LengthUnit toUnit);
+        Length ConvertLength(decimal value, LengthUnit fromUnit, LengthUnit toUnit);
     }
 }
