+++ src/UnitConv/UnitConverter.cs
@@ -6,7 +6,7 @@ namespace UnitConv
 {
     public class UnitConverter : IUnitConverter
     {
-        public Length ConvertLength(float value, LengthUnit fromUnit, LengthUnit toUnit)
+        public Length ConvertLength(decimal value, LengthUnit fromUnit, LengthUnit toUnit)
         {
             return new Length(value, fromUnit).ConvertValue(toUnit);
         }
